load a new random gif only when the app is in foreground
Fixed a major issue causing the app would indefinitely load a new random
GIF in the detail screen every 10 seconds, even when the app was in the
background.

Now the ViewModel checks if there is any active observer first, and then
makes the call for a new GIF, or skips it otherwise. There is no point
in consuming data for something that a user may never see. (At least
until the mobile data provider charges him a ton of gigabytes, because
he had the app running in the background for the past day.)